added utils class
